COMP: Fix warning about minus operator applied to unsigned type
itk::Vector and itk::VariableLengthVector have to be wrapped for the
unsigned int template when wrapping the classes affected by this commit.
Wrapping those two itk vector classes for unsigned int would raise multiple
warnings:

'warning C4146: unary minus operator applied to unsigned type, result
 still unsigned'

Using the int template instead does not affect the logic and prevent
unauthorized wrapping templates to be required.

+++ include/rtkProjectionsDecompositionNegativeLogLikelihood.h
@@ -57,7 +57,7 @@ public:
   typedef vnl_matrix<double>                        MaterialAttenuationsType;
   typedef vnl_matrix<float>                         IncidentSpectrumType;
   typedef itk::VariableLengthVector<double>         MeasuredDataType;
-  typedef itk::VariableLengthVector<unsigned int>   ThresholdsType;
+  typedef itk::VariableLengthVector<int>            ThresholdsType;
   typedef itk::VariableSizeMatrix<double>           MeanAttenuationInBinType;
 
   // Constructor
@@ -160,7 +160,7 @@ public:
       {
       double accumulate = 0;
       double accumulateWeights = 0;
-      for (unsigned int energy=m_Thresholds[bin]-1; (energy<m_Thresholds[bin+1]) && (energy < this->m_MaterialAttenuations.rows()); energy++)
+      for (int energy=m_Thresholds[bin]-1; (energy<m_Thresholds[bin+1]) && (energy < (int)(this->m_MaterialAttenuations.rows())); energy++)
         {
         accumulate += this->m_MaterialAttenuations[energy][mat] * this->m_IncidentSpectrum[0][energy];
         accumulateWeights += this->m_IncidentSpectrum[0][energy];
